Benchmarks: Give both tests option to reformat

+++ vc-benchmarks/increasing_file_size.py
@@ -35,6 +35,9 @@ def parse_args():
     parser.add_argument("--tmp-dir", default="/tmp",
             help="directory in which to create and destroy test repos")
 
+    parser.add_argument("--reformat-partition", default=None,
+            help="reformat this device instead of deleting files one-by-one")
+
     args = parser.parse_args()
     if args.end_mag==-1:
         args.end_mag = args.start_mag+1
@@ -134,7 +137,7 @@ class TrialStats:
         self.gc_iowait = None
 
 
-def run_trial(ts, vcsclass, data_gen, tmpdir="/tmp"):
+def run_trial(ts, vcsclass, data_gen, tmpdir="/tmp", reformat_partition=None):
 
     try:
         repodir = tempfile.mkdtemp(prefix='vcs_benchmark', dir=tmpdir)
@@ -179,7 +182,13 @@ def run_trial(ts, vcsclass, data_gen, tmpdir="/tmp"):
         ts.gc_size = repo.check_total_size()
 
     finally:
-        shutil.rmtree(repodir)
+        log("Cleaning up trial files...")
+        with StopWatch(ts, 'cleanup_time'):
+            if reformat_partition:
+                reformat_device(reformat_partition)
+            else:
+                shutil.rmtree(repodir)
+        log("Removed trial files in %5.3f seconds" % ts.cleanup_time)
 
 
 if __name__ == "__main__":
@@ -199,6 +208,7 @@ if __name__ == "__main__":
             "data_gen": args.data_gen,
             "vcs": args.vcs,
             "vcs_version": vcs_version,
+            "reformat_partition": args.reformat_partition,
         }))
     comment()
     comment(align_kvs(env))
@@ -212,6 +222,13 @@ if __name__ == "__main__":
     printheader(TrialStats.columns)
 
     for mag in range(args.start_mag, args.end_mag):
+
+        # If reformatting, do one at the beginning to ensure all runs start
+        # with the same conditions (last run might have been cancelled without
+        # reformatting)
+        if args.reformat_partition:
+            reformat_device(args.reformat_partition)
+
         for step in range(0, args.mag_steps):
             bytesperstep = 2**mag / args.mag_steps
             filebytes = 2**mag + step*bytesperstep
